Trying to find cause of widely varying performance
